Put a small sleep in the timer function else case
To prevent high CPU usage, adding a small sleep in the else logic.

This should also alleviate some UI lag issues in Openroast once you plug the roaster in. I found that the tight loop in the timer thread was taking so much CPU that the UI responsiveness would drop to about one second for a click on my machine.